+++ controller/USER_Controller.php
@@ -3,6 +3,8 @@ require_once(__DIR__."/../core/ViewManager.php");
 require_once(__DIR__."/../core/I18n.php");
 require_once(__DIR__."/../model/User.php");
 require_once(__DIR__."/../model/UserMapper.php");
+require_once(__DIR__."/../model/Profile.php");
+require_once(__DIR__."/../model/ProfileMapper.php");
 require_once(__DIR__."/../controller/BaseController.php");
 
 class USER_Controller extends BaseController {
@@ -36,6 +38,22 @@ class USER_Controller extends BaseController {
         $this->view->render("user", "show");
     }
 
+    public function showone(){
+        if (!isset($_REQUEST["id"])) {
+            throw new Exception("An user id is mandatory");
+        }
+
+        $userid = $_REQUEST["id"];
+        $user = $this->userMapper->fetch($userid);
+        
+        if ($user == NULL) {
+            throw new Exception("no such user with id: ".$userid);
+        }
+
+        $this->view->setVariable("user", $user);
+        $this->view->render("user", "showone");
+    }
+
     public function insert(){
         //checkPermissionsNeed
     
@@ -47,27 +65,33 @@ class USER_Controller extends BaseController {
             $user->setProfile($_POST["profile"]);
 			 
             try {
-                $user->checkIsValidForCreate();
-                $this->userMapper->insert($user);
-	
-                $this->view->setFlash(sprintf(i18n("User \"%s\" successfully added."),$user->getUsername()));
-	
-                $this->view->redirect("user", "show");
-	
+                if (!$this->userMapper->usernameExists($_POST["username"])){
+                    $user->checkIsValidForCreate();
+                    $this->userMapper->insert($user);
+                    $this->view->setFlash(sprintf(i18n("User \"%s\" successfully added."),$user->getUsername()));
+                    $this->view->redirect("user", "show");
+                } else {
+                    $errors = array();
+	                $errors["general"] = "Username already exists";
+	                $this->view->setVariable("errors", $errors);
+                }
             }catch(ValidationException $ex) {      
-                $errors = $ex->getErrors();	
+                $errors = $ex->getErrors();
                 $this->view->setVariable("errors", $errors);
             }
         }
-    
+
+        $profileMapper = new ProfileMapper();
+        $profiles = $profileMapper->fetch_all();
         $this->view->setVariable("user", $user);
+        $this->view->setVariable("profiles", $profiles);
         $this->view->render("user", "insert");
     }
 
 
     public function update() {
         if (!isset($_REQUEST["id"])) {
-            throw new Exception("A user id is mandatory");
+            throw new Exception("An user id is mandatory");
         }
 
         //CheckPermissionsNeed
@@ -84,17 +108,26 @@ class USER_Controller extends BaseController {
             $user->setProfile($_POST["profile"]);
       
             try {
-                $user->checkIsValidForCreate();
-                $this->userMapper->update($user);                
-                $this->view->setFlash(sprintf(i18n("User \"%s\" successfully updated."),$user->getUsername()));
-                $this->view->redirect("user", "show");	
+                if (!$this->userMapper->usernameExists($user->getUsername())){
+                    $user->checkIsValidForCreate();
+                    $this->userMapper->update($user);                
+                    $this->view->setFlash(sprintf(i18n("User \"%s\" successfully updated."),$user->getUsername()));
+                    $this->view->redirect("user", "show");
+                } else {
+                    $errors = array();
+	                $errors["general"] = "Username already exists";
+	                $this->view->setVariable("errors", $errors);
+                }
             }catch(ValidationException $ex) {
                 $errors = $ex->getErrors();
                 $this->view->setVariable("errors", $errors);
             }
         }
-    
+
+        $profileMapper = new ProfileMapper();
+        $profiles = $profileMapper->fetch_all();
         $this->view->setVariable("user", $user);
+        $this->view->setVariable("profiles", $profiles);
         $this->view->render("user", "update");    
     }
 
@@ -111,10 +144,16 @@ class USER_Controller extends BaseController {
         if ($user == NULL) {
             throw new Exception("no such user with id: ".$userid);
         }
-    
-        $this->userMapper->delete($user);
-        $this->view->setFlash(sprintf(i18n("User \"%s\" successfully deleted."),$user->getUsername()));
-        $this->view->redirect("user", "show");
+
+        if (isset($_POST["submit"])) {
+            if ($_POST["submit"] == "yes"){
+                $this->userMapper->delete($user);
+                $this->view->setFlash(sprintf(i18n("User \"%s\" successfully deleted."),$user->getUsername()));
+            }
+            $this->view->redirect("user", "show");
+        }
+        $this->view->setVariable("user", $user);
+        $this->view->render("user", "delete");
     }
 
     public function logout() {
